requirements.txt
# Knowledge Graph Requirements
neo4j==5.12.0  # Graph database
py2neo==2021.2.3  # Neo4j Python driver
networkx==3.1  # Graph manipulation

# AI and ML Requirements
openai==1.5.0  # OpenAI API for LLM
langchain==0.0.313  # Framework for LLM applications
langchain-openai==0.0.2  # OpenAI integration for langchain
sentence-transformers==2.2.2  # Embeddings
transformers==4.35.2  # Hugging Face transformers
torch==2.1.1  # PyTorch for ML operations
scikit-learn==1.3.2  # ML utilities

# Document Processing
pypdf==3.17.0  # PDF processing
pytesseract==0.3.10  # OCR capabilities
pdf2image==1.16.3  # PDF to image conversion
pillow==10.1.0  # Image processing
pandas==2.1.3  # Data manipulation
openpyxl==3.1.2  # Excel file processing
python-docx==1.0.1  # Word document processing

# Web Framework and API
flask==2.3.3  # Web framework
flask-restx==1.1.0  # REST API framework for Flask
flask-cors==4.0.0  # Cross-Origin Resource Sharing
flask-jwt-extended==4.5.3  # JWT for authentication
gunicorn==21.2.0  # WSGI HTTP Server

# Database
sqlalchemy==2.0.23  # ORM
alembic==1.12.1  # Database migrations
psycopg2-binary==2.9.9  # PostgreSQL driver

# Security
cryptography==41.0.5  # Cryptographic recipes
bcrypt==4.0.1  # Password hashing
python-dotenv==1.0.0  # Environment variable management

# Testing
pytest==7.4.3  # Testing framework
pytest-cov==4.1.0  # Test coverage
requests-mock==1.11.0  # Mock HTTP requests

# Utilities
tqdm==4.66.1  # Progress bars
pydantic==2.5.2  # Data validation
python-dateutil==2.8.2  # Date utilities
jsonschema==4.20.0  # JSON schema validation 